requirements.txt
dvc
pandas
numpy
biopython
matplotlib
seaborn
